Changed transaction table to functional component
