+++ package.json
@@ -15,7 +15,8 @@
     "react": "^16.13.1",
     "react-datepicker": "^2.14.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "uuid": "^8.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
