+++ src/App.js
@@ -2,6 +2,7 @@ import React from "react";
 import Transactions from "./components/Transactions";
 import AddTransaction from "./components/AddTransaction";
 import axios from "axios";
+import { v4 as uuidv4 } from "uuid";
 
 const apiUrl = "http://localhost:9000/transactions";
 
@@ -10,27 +11,69 @@ class App extends React.Component {
     super();
 
     this.state = {
+      transactions: [],
       showAddTransaction: false,
-      id: 1
+      id: 1,
+      uuid: 0
     };
 
     this.onAddButtonClick = this.onAddButtonClick.bind(this);
     this.onAddTransaction = this.onAddTransaction.bind(this);
+    this.getTransactions = this.getTransactions.bind(this);
+    this.onDeleteTransaction = this.onDeleteTransaction.bind(this);
   }
 
+  componentDidMount() {
+    this.getTransactions();
+  }
+
+  // Get transactions from the API
+  getTransactions() {
+    fetch(apiUrl)
+      .then(response => response.json())
+      .then(data =>
+        this.setState({
+          transactions: data
+        })
+      )
+      .catch(function (err) {
+        console.log(err);
+      });
+  }
+
+  // Show add transaction modal when clicked
   onAddButtonClick() {
     this.setState(prevState => ({
       showAddTransaction: !prevState.showAddTransaction
     }));
   }
 
-  onAddTransaction(transaction) {
-    axios.post(apiUrl, transaction).then(
+  // Post the new transaction, get transactions from API to update table
+  async onAddTransaction(transaction) {
+    await axios.post(apiUrl, transaction).then(
       this.setState(prevState => ({
         showAddTransaction: !prevState.showAddTransaction,
-        id: prevState.id + 1
+        uuid: uuidv4()
       }))
     );
+    this.getTransactions();
+  }
+
+  onDeleteTransaction(id, i) {
+    // Delete transaction by id
+    fetch(`${apiUrl}/${id}`, {
+      method: "delete"
+    })
+      .then(response => response.json())
+      .then(() => {
+        var newTransactions = [...this.state.transactions]; // Create copy of transactions
+        newTransactions.splice(i, 1); // Remove deleted transaction
+        this.setState({ transactions: newTransactions }); // Set state to new transactions
+      })
+      .catch(err => {
+        console.log(err);
+      });
+    this.getTransactions();
   }
 
   render() {
@@ -42,7 +85,10 @@ class App extends React.Component {
         ) : (
           <button onClick={this.onAddButtonClick}>Add Transaction</button>
         )}
-        <Transactions key={this.state.id} />
+        <Transactions
+          transactions={this.state.transactions}
+          onDelete={this.onDeleteTransaction}
+        />
       </div>
     );
   }
